requirements.txt
python-telegram-bot==20.7
simpleeval
nest_asyncio
